# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TaskTips Cloud API
  version: 0.1.0
  description: Contract-first API for TaskTips synchronization and administration.
  license:
    name: Proprietary
    identifier: LicenseRef-TaskTips-Proprietary
servers:
  - url: /
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Authenticate a device
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access credentials for a regular user account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '408':
          $ref: '#/components/responses/RequestTimeout'
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: Rotate a refresh token
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Rotated access credentials for a regular user account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '408':
          $ref: '#/components/responses/RequestTimeout'
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke all refresh tokens for the authenticated device
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: The authenticated device sessions were revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/auth/login:
    post:
      operationId: adminLogin
      summary: Authenticate a system administrator
      tags: [Administration]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Administrator access token and a Secure HttpOnly tasktips_admin_refresh cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTokenResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/admin/auth/refresh:
    post:
      operationId: adminRefreshToken
      summary: Rotate the administrator refresh cookie
      tags: [Administration]
      security: []
      responses:
        '200':
          description: Rotated administrator access token and refresh cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTokenResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/admin/auth/logout:
    post:
      operationId: adminLogout
      summary: Revoke the administrator device session
      tags: [Administration]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: The administrator device session was revoked.
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/auth/re-auth:
    post:
      operationId: adminReauthenticate
      summary: Re-authenticate an administrator for a sensitive operation
      tags: [Administration]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminReauthRequest'
      responses:
        '200':
          description: A short-lived, one-use operation nonce.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReauthResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/auth/invitations/activate:
    post:
      operationId: activateInvitation
      summary: Activate an invited account
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationActivationRequest'
      responses:
        '200':
          description: The invited account was activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '408':
          $ref: '#/components/responses/RequestTimeout'
  /api/v1/me:
    get:
      operationId: getCurrentUser
      summary: Get the current account metadata
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current account metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/me/password:
    patch:
      operationId: changePassword
      summary: Change the current account password
      tags: [Authentication]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204':
          description: The password was changed.
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List projects owned by the current account
      tags: [Projects]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Projects visible through owner RLS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createProject
      summary: Create a personal project
      tags: [Projects]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Project created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getProject
      summary: Get an owned project
      tags: [Projects]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owned project metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
    patch:
      operationId: updateProject
      summary: Rename an owned project
      tags: [Projects]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: Updated project metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
  /api/v1/projects/{projectId}/disable:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: disableProject
      summary: Disable an owned project without deleting history
      tags: [Projects]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Disabled project metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
  /api/v1/projects/{projectId}/purge:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: purgeProject
      summary: Verify the owner password and queue irreversible project deletion
      tags: [Projects]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurgeProjectRequest'
      responses:
        '202':
          description: Project deletion queued. The project enters deleting state immediately.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeJob'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/projects/{projectId}/sync/bootstrap:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: bootstrapSync
      summary: Read all heads and tombstones from a stable change sequence
      tags: [Synchronization]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
      responses:
        '200':
          description: A stable bootstrap page. cursor is present only on the final page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
        '409':
          $ref: '#/components/responses/GenerationMismatch'
        '423':
          $ref: '#/components/responses/ProjectMaintenance'
  /api/v1/projects/{projectId}/sync/pull:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: pullSyncChanges
      summary: Pull permanent change-log entries after a signed cursor
      tags: [Synchronization]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PullRequest'
      responses:
        '200':
          description: Incremental object and tombstone changes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullResponse'
        '400':
          $ref: '#/components/responses/CursorInvalid'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
        '409':
          $ref: '#/components/responses/GenerationMismatch'
        '423':
          $ref: '#/components/responses/ProjectMaintenance'
  /api/v1/projects/{projectId}/sync/push:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: pushSyncChanges
      summary: Push up to 100 object and tombstone CAS changes
      tags: [Synchronization]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRequest'
      responses:
        '200':
          description: Per-item applied, conflict, or rejected results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
        '409':
          description: Generation, bootstrap, or idempotency request-level rejection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          $ref: '#/components/responses/ProjectMaintenance'
...
  /api/v1/devices:
    get:
      operationId: listDevices
      summary: List devices owned by the current account
      tags: [Devices]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Devices visible through owner RLS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InvalidRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the credentials are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The account or device is disabled, revoked, or not authorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: The authentication endpoint rate limit was exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password, deviceId]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
        deviceId:
          type: string
          format: uuid
...
    TokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          minimum: 1
...